+++ app/monthly-cleaning/page.tsx
@@ -1,0 +1,5 @@
+import MonthlyCleaningSheet from "../sheets/monthly-cleaning/MonthlyCleaningSheet";
+
+export default function MonthlyCleaningPage() {
+  return <MonthlyCleaningSheet />;
+}